😱

+++ publish.ps1
@@ -11,33 +11,38 @@ if ($api_key.Length -eq 0)
     exit
 }
 
-$projects = "src/libraries/HLE", "src/libraries/HLE.Twitch"
-
-foreach ($project in $projects)
+$libraries_directory = "src/libraries"
+if ($args.Length -eq 0 -or $args[0].Length -eq 0)
 {
-    Push-Location
-    try
+    Write-Error "No library specified. Please provide the library name as an argument."
+    exit
+}
+
+$library = $args[0]
+$library_directory = [System.IO.Path]::Combine($libraries_directory, $library)
+
+Push-Location
+try
+{
+    Set-Location $library_directory
+    Remove-Item "bin/Release/*.nupkg"
+
+    dotnet pack -c Release -p:PublishingPackage=true
+    if ($LastExitCode -ne 0)
     {
-        Set-Location "$project"
-        Remove-Item "bin/Release/*.nupkg"
-
-        dotnet pack -c Release -p:PublishingPackage=true
-        if ($LastExitCode -ne 0)
-        {
-            exit
-        }
-
-        Set-Location "bin/Release"
-        $nuget_package = Get-Item "*.nupkg"
-
-        dotnet nuget push $nuget_package --api-key $api_key --source $nuget_source --skip-duplicate
-        if ($LastExitCode -ne 0)
-        {
-            exit
-        }
+        exit
     }
-    finally
+
+    Set-Location "bin/Release"
+    $nuget_package = Get-Item "*.nupkg"
+
+    dotnet nuget push $nuget_package --api-key $api_key --source $nuget_source --skip-duplicate
+    if ($LastExitCode -ne 0)
     {
-        Pop-Location
+        exit
     }
+}
+finally
+{
+    Pop-Location
 }
